# services/academic-core-service/docs/openapi-front-office.yaml
...
openapi: 3.0.3
info:
  title: Academic Core Service – Front Office API
  description: |
    API for managing front office operations within the academic-core-service.
    This module handles day-to-day administrative activities including:
    - Visitors Log
    - Phone Call Log
    - Half Day Notices
    - Postal Dispatch/Receive
    - Admission Enquiries
    - Complaints
    
    All operations are scoped to a specific school and academic year.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Base API path
...
paths:
...
  /front-office/admission-enquiries:
    get:
      summary: List admission enquiries
      description: |
        Retrieve a paginated list of admission enquiries.
        Supports filtering by status, source, date, and follow-up tracking.
      tags: [Admission Enquiries]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - $ref: '#/components/parameters/AcademicYearIdHeader'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/EnquiryStatusEnum'
          description: Filter by enquiry status
        - name: source
          in: query
          schema:
            $ref: '#/components/schemas/EnquirySourceEnum'
          description: Filter by enquiry source
        - name: enquiryType
          in: query
          schema:
            $ref: '#/components/schemas/EnquiryTypeEnum'
          description: Filter by enquirer type
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
          description: Filter enquiries from this date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
          description: Filter enquiries until this date
        - name: hasFollowUp
          in: query
          schema:
            type: boolean
          description: Filter enquiries with pending follow-up
        - name: search
          in: query
          schema:
            type: string
          description: Search by enquirer name or phone number
      responses:
        '200':
          description: Paginated list of admission enquiries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionEnquiryPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create admission enquiry
      description: Record a new admission enquiry.
      tags: [Admission Enquiries]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - $ref: '#/components/parameters/AcademicYearIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdmissionEnquiryRequest'
      responses:
        '201':
          description: Admission enquiry created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionEnquiryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /front-office/admission-enquiries/{id}:
    get:
      summary: Get admission enquiry details
      description: Retrieve details of a specific admission enquiry.
      tags: [Admission Enquiries]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Admission enquiry details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionEnquiryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /front-office/admission-enquiries/{id}/status:
    patch:
      summary: Update admission enquiry status
      description: Update the status of an admission enquiry.
      tags: [Admission Enquiries]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEnquiryStatusRequest'
      responses:
        '200':
          description: Enquiry status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionEnquiryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /front-office/admission-enquiries/{id}/follow-up:
    patch:
      summary: Update admission enquiry follow-up
      description: Update the follow-up dates of an admission enquiry.
      tags: [Admission Enquiries]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEnquiryFollowUpRequest'
      responses:
        '200':
          description: Follow-up dates updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionEnquiryResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # -------------------------------------------------------------------------
  # COMPLAINTS
  # -------------------------------------------------------------------------
  /front-office/complaints:
    get:
      summary: List complaints
      description: |
        Retrieve a paginated list of complaints.
        Supports filtering by status, complaint type, category, and assigned staff.
      tags: [Complaints]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - $ref: '#/components/parameters/AcademicYearIdHeader'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ComplaintStatusEnum'
          description: Filter by complaint status
        - name: complaintType
          in: query
          schema:
            $ref: '#/components/schemas/ComplaintTypeEnum'
          description: Filter by complaint type
        - name: category
          in: query
          schema:
            type: string
          description: Filter by complaint category
        - name: assignedToStaffId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by assigned staff member
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
          description: Filter complaints from this date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
          description: Filter complaints until this date
        - name: search
          in: query
          schema:
            type: string
          description: Search by complainant name
      responses:
        '200':
          description: Paginated list of complaints
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: File complaint
      description: Record a new complaint.
      tags: [Complaints]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - $ref: '#/components/parameters/AcademicYearIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComplaintRequest'
      responses:
        '201':
          description: Complaint filed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /front-office/complaints/{id}:
    get:
      summary: Get complaint details
      description: Retrieve details of a specific complaint.
      tags: [Complaints]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Complaint details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /front-office/complaints/{id}/assign:
    patch:
      summary: Assign complaint to staff
      description: Assign a complaint to a staff member for resolution.
      tags: [Complaints]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignComplaintRequest'
      responses:
        '200':
          description: Complaint assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /front-office/complaints/{id}/status:
    patch:
      summary: Update complaint status
      description: Update the status and action taken for a complaint.
      tags: [Complaints]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateComplaintStatusRequest'
      responses:
        '200':
          description: Complaint status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplaintResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

# =============================================================================
# COMPONENTS
# =============================================================================
components:
...
  parameters:
    SchoolIdHeader:
      name: X-School-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: The school context for the request
    AcademicYearIdHeader:
      name: X-Academic-Year-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: The academic year context for the request
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Page number (0-indexed)
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: Sort criteria (e.g., 'createdAt,desc')

  # ---------------------------------------------------------------------------
  # RESPONSES
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Bad request - validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict - operation not allowed in current state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------------
  # SCHEMAS
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
    PageMetadata:
      type: object
      properties:
        page:
          type: integer
          description: Current page number (0-indexed)
        size:
          type: integer
          description: Number of items per page
        totalElements:
          type: integer
          format: int64
          description: Total number of elements
        totalPages:
          type: integer
          description: Total number of pages

    # --- Enums ---
...
    EnquiryStatusEnum:
      type: string
      enum: [NEW, FOLLOW_UP, CONVERTED, CLOSED]
      description: Status of admission enquiry
    EnquirySourceEnum:
      type: string
      enum: [WEBSITE, WALK_IN, CALL, REFERRAL, OTHER]
      description: Source of admission enquiry
    EnquiryTypeEnum:
      type: string
      enum: [PARENT, STUDENT, OTHER]
      description: Type of enquirer
    ComplaintStatusEnum:
      type: string
      enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
      description: Status of complaint
    ComplaintTypeEnum:
      type: string
      enum: [PARENT, STUDENT, STAFF]
      description: Type of complainant

    # --- Visitors ---
...
    AdmissionEnquiryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        enquirerName:
          type: string
        phoneNumber:
          type: string
        enquiryType:
          $ref: '#/components/schemas/EnquiryTypeEnum'
        source:
          $ref: '#/components/schemas/EnquirySourceEnum'
        enquiryDate:
          type: string
          format: date
        description:
          type: string
          nullable: true
        lastFollowUpDate:
          type: string
          format: date
          nullable: true
        nextFollowUpDate:
          type: string
          format: date
          nullable: true
        status:
          $ref: '#/components/schemas/EnquiryStatusEnum'
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
    CreateAdmissionEnquiryRequest:
      type: object
      required: [enquirerName, phoneNumber, enquiryType, source, enquiryDate]
      properties:
        enquirerName:
          type: string
          minLength: 2
          maxLength: 100
        phoneNumber:
          type: string
          pattern: '^\+?[0-9]{10,15}$'
        enquiryType:
          $ref: '#/components/schemas/EnquiryTypeEnum'
        source:
          $ref: '#/components/schemas/EnquirySourceEnum'
        enquiryDate:
          type: string
          format: date
        description:
          type: string
          maxLength: 1000
        nextFollowUpDate:
          type: string
          format: date
        remarks:
          type: string
          maxLength: 500
    UpdateEnquiryStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/EnquiryStatusEnum'
        remarks:
          type: string
          maxLength: 500
    UpdateEnquiryFollowUpRequest:
      type: object
      properties:
        lastFollowUpDate:
          type: string
          format: date
        nextFollowUpDate:
          type: string
          format: date
        remarks:
          type: string
          maxLength: 500
    AdmissionEnquiryPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AdmissionEnquiryResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'

    # --- Complaints ---
    ComplaintResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        complainantName:
          type: string
        complaintType:
          $ref: '#/components/schemas/ComplaintTypeEnum'
        category:
          type: string
        complaintDate:
          type: string
          format: date
        description:
          type: string
        actionTaken:
          type: string
          nullable: true
        assignedToStaffId:
          type: string
          format: uuid
          nullable: true
        assignedToStaffName:
          type: string
          nullable: true
          description: Denormalized for display
        internalNote:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ComplaintStatusEnum'
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
    CreateComplaintRequest:
      type: object
      required: [complainantName, complaintType, category, complaintDate, description]
      properties:
        complainantName:
          type: string
          minLength: 2
          maxLength: 100
        complaintType:
          $ref: '#/components/schemas/ComplaintTypeEnum'
        category:
          type: string
          minLength: 2
          maxLength: 100
        complaintDate:
          type: string
          format: date
        description:
          type: string
          minLength: 10
          maxLength: 2000
        internalNote:
          type: string
          maxLength: 1000
        remarks:
          type: string
          maxLength: 500
    AssignComplaintRequest:
      type: object
      required: [assignedToStaffId]
      properties:
        assignedToStaffId:
          type: string
          format: uuid
        internalNote:
          type: string
          maxLength: 1000
    UpdateComplaintStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/ComplaintStatusEnum'
        actionTaken:
          type: string
          maxLength: 2000
        internalNote:
          type: string
          maxLength: 1000
        remarks:
          type: string
          maxLength: 500
    ComplaintPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ComplaintResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'